requirements.txt
alembic==1.7.7
amqp==5.1.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
bcrypt==3.2.0
billiard==3.6.4.0
black==22.3.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-resources==5.6.0
iniconfig==1.1.1
kombu==5.2.4
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.8
pytest==7.1.1
python-decouple==3.6
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
redis==4.2.2
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.8.0
